FiberNlse API

FiberNlse.dispersion — Method

dispersion(D::Float64, λ::Float64)

Generates the Dispersion structure for a fiber
- D : dispersion factor in s/m²
- λ : central wavelength of the study in m

FiberNlse.dispersion — Method

dispersion(β::Float64)

Generates the Dispersion structure for a fiber
- β : second derivative of the propagation constants

FiberNlse.dispersion — Method

dispersion(β::Vector{Float64})

Generates the Dispersion structure for a fiber
- β : nth derivative of the propagation constants from 2 to n

FiberNlse.edfa — Method

edfa(L::Float64, G::Float64, λ::Float64)

Generates a fiber of length L [m] and gain G [dB] arround a wavelength of λ [m]

See also : (smf28)[@ref smf28]

FiberNlse.output — Method

output(f::Field)

returns the signal at the end of the propagation

FiberNlse.smf28 — Method

smf28(L::Float64, λ::Float64)

Generates a standard monomode fiber (smf28) of length L [m] arround a wavelength of λ [m]

See also : (edfa)[@ref edfa]

FiberNlse.Dispersion — Type

Dispersion(β::Union{Float64,Vector{Float64}})

Structure containing dispersion caracteristics for the fiber

Can contain the dispersion only (second derivative of the propagation constant) or a list of slopes from second order to nth order

FiberNlse.Fiber — Type

Fiber(L::Float64, D::Dispersion, γ::Float64, α::Float64)

Structure containing fiber caracteristics :
- L : length in [m]
- D : (dispersion)[@ref Dispersion]
- γ : nonlinear factor in [1 / W / m]
- α : fiber attenuation or gain (exponential) in [1 / m]

See also : (smf28)[@ref smf28], (edfa)[@ref edfa]

FiberNlse.Field — Type

Field(ψ::Matrix{ComplexF64}, l::Vector{Float64}, t::Vector{Float64})

Structure containing space/time data resulting from the simulation
- ψ : The signal propagation matrix
- l : The space vector in m
- t : The time vector in s